requirements.txt
anyio==3.2.1
appdirs==1.4.4
asgiref==3.4.1
Authlib==0.15.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
drf-jwt==1.19.1
filelock==3.0.12
green==3.2.6
h11==0.12.0
httpcore==0.13.6
httpx==0.18.1
idna==3.2
lxml==4.6.3
postgres==3.0.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2018.7
redgreenunittest==0.1.1
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
tda-api==1.3.4
toml==0.10.2
Unidecode==1.2.0
urllib3==1.26.6
wcwidth==0.2.5
websockets==9.1
whitenoise==5.3.0
